+++ src/main/java/modernmods/modernfoundry/tools/data/material/MaterialStatsDataProvider.java
@@ -569,6 +569,9 @@ public class MaterialStatsDataProvider extends AbstractMaterialStatsDataProvider
     addMaterialStats(MaterialIds.necroticBone, RepairStats.ribcage(125));
     addMaterialStats(MaterialIds.necronium,    RepairStats.ribcage(357));
     addMaterialStats(MaterialIds.blaze,        RepairStats.ribcage(150));
+    addMaterialStats(MaterialIds.blazingBone,  RepairStats.ribcage(205));
+    addMaterialStats(MaterialIds.jadeite,      RepairStats.ribcage(616));
+    addMaterialStats(MaterialIds.endRod,       RepairStats.ribcage(250));
 
     // slimeshell
     addMaterialStats(MaterialIds.turtle, RepairStats.shell(150));